Refactored pattern structure: Renamed 'allPattern' to 'PatternGrid' and created 'allPatternPage' to reuse 'PatternGrid' in 'ProfilePage'. This avoids code duplication, as 'PatternGrid' is identical in both pages, except for different API calls (all patterns vs. user-liked patterns).
