+++ frontend/src/pages/DashboardPage.tsx
@@ -1,13 +1,13 @@
 import { useContext } from "react";
 import AboutPage from "./AboutPage";
-import AllPattern from "../components/AllPattern";
+import AllPatternsPage from "./AllPatternsPage";
 import { AuthContext } from "../context/auth.context";
 
 const DashboardPage: React.FC = () => {
   const authContext = useContext(AuthContext);
   const isLoggedIn = authContext ? authContext.isLoggedIn : false;
 
-  return <div>{isLoggedIn ? <AllPattern /> : <AboutPage />}</div>;
+  return <div>{isLoggedIn ? <AllPatternsPage /> : <AboutPage />}</div>;
 };
 
 export default DashboardPage;
